Review type added for review post
