+++ src/testing/mocks/data/reviews.mock.ts
@@ -1,0 +1,5 @@
+export const mockReviewPost = {
+  assignment_id: 12345,
+  incorrect_meaning_answers: 2,
+  incorrect_reading_answers: 1,
+};
